+++ src/StorageBase/CHANGELOG.md
@@ -41,6 +41,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Change string parameter in `\Heptacom\HeptaConnect\Storage\Base\Action\PortalNode\Get\PortalNodeGetResult::__construct`, `\Heptacom\HeptaConnect\Storage\Base\Action\PortalNode\Overview\PortalNodeOverviewResult::__construct`, `\Heptacom\HeptaConnect\Storage\Base\Action\PortalExtension\Find\PortalExtensionFindResult::add` and `\Heptacom\HeptaConnect\Storage\Base\Action\Route\Overview\RouteOverviewResult::__construct` to be a `\Heptacom\HeptaConnect\Dataset\Base\Contract\ClassStringReferenceContract` instead of a string for better [type safe class strings](https://heptaconnect.io/reference/adr/2022-06-12-type-safe-class-strings/)
 - Change array of string parameter in `\Heptacom\HeptaConnect\Storage\Base\Action\PortalNode\Overview\PortalNodeOverviewCriteria::setClassNameFilter` to be a `\Heptacom\HeptaConnect\Dataset\Base\Support\ClassStringReferenceCollection` instead of a string for better [type safe class strings](https://heptaconnect.io/reference/adr/2022-06-12-type-safe-class-strings/)
 - Implement `\Heptacom\HeptaConnect\Dataset\Base\Contract\AttachmentAwareInterface` in `\Heptacom\HeptaConnect\Storage\Base\PrimaryKeySharingMappingStruct`
+- Change type of capabilities on `\Heptacom\HeptaConnect\Storage\Base\Action\Route\Overview\RouteOverviewResult` from untyped array to `\Heptacom\HeptaConnect\Dataset\Base\ScalarCollection\StringCollection`
 
 ### Deprecated